refactor(app): remove deprecated cardlist
