+++ app/__tests__/components/CardList.test.tsx
@@ -18,27 +18,27 @@ const mockCeramicContext: CeramicContextState = makeTestCeramicContext();
 let cardListProps: CardListProps = {};
 
 describe("<CardList />", () => {
-  // beforeEach(() => {
-  //   cardListProps = {};
-  // });
-
-  it("renders provider cards when loading state is not defined", () => {
-    // renderWithContext(mockUserContext, mockCeramicContext, <CardList {...cardListProps} />);
-    // expect(screen.queryByTestId("loading-card")).not.toBeInTheDocument();
+  beforeEach(() => {
+    cardListProps = {};
   });
 
-  // it("renders provider cards when not loading", () => {
-  //   cardListProps.isLoading = false;
-  //   renderWithContext(mockUserContext, mockCeramicContext, <CardList {...cardListProps} />);
+  it("renders provider cards when loading state is not defined", () => {
+    renderWithContext(mockUserContext, mockCeramicContext, <CardList {...cardListProps} />);
+    expect(screen.queryByTestId("loading-card")).not.toBeInTheDocument();
+  });
 
-  //   expect(screen.queryByTestId("loading-card")).not.toBeInTheDocument();
-  // });
+  it("renders provider cards when not loading", () => {
+    cardListProps.isLoading = false;
+    renderWithContext(mockUserContext, mockCeramicContext, <CardList {...cardListProps} />);
 
-  // it("renders LoadingCards when loading the passport", () => {
-  //   cardListProps.isLoading = true;
+    expect(screen.queryByTestId("loading-card")).not.toBeInTheDocument();
+  });
 
-  //   renderWithContext(mockUserContext, mockCeramicContext, <CardList {...cardListProps} />);
+  it("renders LoadingCards when loading the passport", () => {
+    cardListProps.isLoading = true;
 
-  //   expect(screen.getAllByTestId("loading-card"));
-  // });
+    renderWithContext(mockUserContext, mockCeramicContext, <CardList {...cardListProps} />);
+
+    expect(screen.getAllByTestId("loading-card"));
+  });
 });
